Updated version intended to work with Python3.11
Existing version of wax coppied to wax2 (for python2)
wax directory contains local version intended to work Python3

Dependencies for this verson of code are in requirements.txt
Code should be run from a venv environment setup with this
requirements.txt file

Note that due to bugs in the latest version of wxpython,
we are currently using 4.1.1 and a downgrade of numpy that is
compatible.

The dependency on ply seems to be no longer relevant
(but the code remains for now).

The dependency on grandalf is now met through venv/pip
(but the code remains for now).

All mwp classes have been upgraded to Python3

Config files for VSCode added
(but the Eclipse files remain for now)